+++ linkis-public-enhancements/linkis-datasource/linkis-metadata-query/service/hive/src/main/java/org/apache/linkis/metadata/query/service/HiveParamsMapper.java
@@ -15,24 +15,21 @@
  * limitations under the License.
  */
 
-package org.apache.linkis.metadat.query.service;
+package org.apache.linkis.metadata.query.service;
 
 import org.apache.linkis.common.conf.CommonVars;
 
-public class SqlParamsMapper {
+public class HiveParamsMapper {
 
-    public static final CommonVars<String> PARAM_SQL_HOST =
-            CommonVars.apply("wds.linkis.server.mdm.service.sql.host", "host");
+    public static final CommonVars<String> PARAM_HIVE_PRINCIPLE =
+            CommonVars.apply("wds.linkis.server.mdm.service.hive.principle", "principle");
 
-    public static final CommonVars<String> PARAM_SQL_PORT =
-            CommonVars.apply("wds.linkis.server.mdm.service.sql.port", "port");
+    public static final CommonVars<String> PARAM_HIVE_URIS =
+            CommonVars.apply("wds.linkis.server.mdm.service.hive.uris", "uris");
 
-    public static final CommonVars<String> PARAM_SQL_USERNAME =
-            CommonVars.apply("wds.linkis.server.mdm.service.sql.username", "username");
+    public static final CommonVars<String> PARAM_HIVE_KEYTAB =
+            CommonVars.apply("wds.linkis.server.mdm.service.hive.keytab", "keytab");
 
-    public static final CommonVars<String> PARAM_SQL_PASSWORD =
-            CommonVars.apply("wds.linkis.server.mdm.service.sql.password", "password");
-
-    public static final CommonVars<String> PARAM_SQL_EXTRA_PARAMS =
-            CommonVars.apply("wds.linkis.server.mdm.service.sql.params", "params");
+    public static final CommonVars<String> PARAM_HADOOP_CONF =
+            CommonVars.apply("wds.linkis.server.mdm.service.hadoop.conf", "hadoopConf");
 }
