+++ linkis-public-enhancements/linkis-datasource/linkis-metadata-query/service/kafka/src/main/java/org/apache/linkis/metadata/query/service/KafkaParamsMapper.java
@@ -15,21 +15,17 @@
  * limitations under the License.
  */
 
-package org.apache.linkis.metadat.query.service;
+package org.apache.linkis.metadata.query.service;
 
 import org.apache.linkis.common.conf.CommonVars;
 
-public class HiveParamsMapper {
+public class KafkaParamsMapper {
+    public static final CommonVars<String> PARAM_KAFKA_PRINCIPLE =
+            CommonVars.apply("wds.linkis.server.mdm.service.kafka.principle", "principle");
 
-    public static final CommonVars<String> PARAM_HIVE_PRINCIPLE =
-            CommonVars.apply("wds.linkis.server.mdm.service.hive.principle", "principle");
+    public static final CommonVars<String> PARAM_KAFKA_KEYTAB =
+            CommonVars.apply("wds.linkis.server.mdm.service.kafka.keytab", "keytab");
 
-    public static final CommonVars<String> PARAM_HIVE_URIS =
-            CommonVars.apply("wds.linkis.server.mdm.service.hive.uris", "uris");
-
-    public static final CommonVars<String> PARAM_HIVE_KEYTAB =
-            CommonVars.apply("wds.linkis.server.mdm.service.hive.keytab", "keytab");
-
-    public static final CommonVars<String> PARAM_HADOOP_CONF =
-            CommonVars.apply("wds.linkis.server.mdm.service.hadoop.conf", "hadoopConf");
+    public static final CommonVars<String> PARAM_KAFKA_BROKERS =
+            CommonVars.apply("wds.linkis.server.mdm.service.kafka.brokers", "brokers");
 }
